Improve logic controlling the enabling of testing.
When range-v3 is being used as a standalone project (i.e., not a
subproject) then behavior should stay the same, in that the building
of tests/examples/etc. is enabled by default, but still settable
by the user using the same variables.

However, if range-v3 is a subproject of a parent then we will attempt
to respect the parent's settings regarding testing.  In that situation:

  * Testing/header checks are disabled by default.
  * Examples/docs/perf targets are disabled by default.
  * Enable testing/header checks if the parent has already enabled
    testing.

Effectively this causes the range-v3 subproject to mirror the testing
status of its parent, but while still allowing customization using
the usual variables.

This change was tested with CMake 3.13 in the following configurations:

  1) range-v3 is a standalone project, verified that tests/examples
     are built by default.
  2) range-v3 is a subproject of a parent with testing disabled.
     Verified that range-v3 tests are not built.
  3) rnage-v3 is a subproject of a parent with testing enabled.
     Verified that range-v3 tests are built and runnable as part
     of the global collection of tests.